Changing rule which defines duplicate participants in participant-import
Rule use to define a pre-recruit as a duplicate if the pre-recruit has the
same first and last name and a matching phone number to an existing participant.
Now  they are defined as a pre-recruit if the pre-recruit has a phone number
(home or cell) that matches an existing participant's phone number who belongs
to the same cohort.